Learn Julia

This documentation is a record of my notes and findings while learning the Julia programming language and supporting ecosystem. There are certainly better places to find this information (namely the official docs for Julia and corresponding packages).

LearnJulia.ExampleSubmodule

LearnJulia.helloworld

LearnJulia.helloworld — Method

helloworld()

A dreadfully important function which does a thing.

LearnJulia.ExampleSubmodule — Module

This module is an example of how to create and import a submodule.

Functions can be exported using export and are then usable when the parent module has a line like using .ExampleSubmodule.